requirements.txt
# 兼容 Python 3.10+
# 执行安装：pip install -r requirements.txt
apscheduler==3.11.2
pymysql==1.1.2
python-dotenv==1.2.2
requests==2.32.5
schedule==1.2.2
scrapy==2.14.1
selenium==4.41.0
